Добавлена авторизация с комнатными ролями + доработки

+++ FinAnalyzer.Web/Controllers/RoomController.cs
@@ -13,10 +13,12 @@ namespace FinAnalyzer.Web.Controllers;
 public class RoomController : ControllerBase
 {
     private readonly IRoomService _roomService;
+    private readonly IAuthService _authService;
 
-    public RoomController(IRoomService roomService)
+    public RoomController(IRoomService roomService, IAuthService authService)
     {
         _roomService = roomService;
+        _authService = authService;
     }
 
     /// <summary>
@@ -54,11 +56,8 @@ public class RoomController : ControllerBase
     /// <param name="id">id комнаты</param>
     /// <returns></returns>
     [HttpGet("{id}")]
-    public async Task<ActionResult<OperationResult<RoomResponse>>> 
-        GetById(int id)
+    public async Task<ActionResult<OperationResult<RoomResponse>>> GetById(int id)
     {
-        
-
         var result = await _roomService.GetByIdAsync(id);
 
         if (result.Success)
@@ -69,19 +68,24 @@ public class RoomController : ControllerBase
 
         return BadRequest(result);
     }
-    
+
     /// <summary>
     /// Получения коллекции комнат, в которых состоит пользователь
     /// </summary>
-    /// <param name="id">id пользователя</param>
+    /// <param name="personId">id пользователя</param>
     /// <param name="searchText">Поиск по названию и описанию</param>
     /// <param name="skip"></param>
     /// <param name="take"></param>
     /// <returns></returns>
-    [HttpGet("get-by-personid/{id}")]
+    [HttpGet("get-by-personid/{personId}")]
     public async Task<ActionResult<OperationResult<PaginationResponse<RoomResponse>>>>
-        GetByPersonId(int id, string? searchText, int skip = 0, int take = 20)
+        GetByPersonId(int personId, string? searchText, int skip = 0, int take = 20)
     {
+        var jwt = Request.Headers["Authorization"];
+
+        if (!_authService.IsPersonId(jwt, personId))
+            return Forbid();
+
         var pagination = new PaginationRequest
         {
             SearchText = searchText,
@@ -89,7 +93,7 @@ public class RoomController : ControllerBase
             Take = take
         };
 
-        var result = await _roomService.GetByPersonIdAsync(id, pagination);
+        var result = await _roomService.GetByPersonIdAsync(personId, pagination);
 
         if (result.Success)
             return Ok(result);
@@ -108,6 +112,11 @@ public class RoomController : ControllerBase
     [HttpPost("create")]
     public async Task<ActionResult<OperationResult<int>>> Create(RoomCreateRequest request)
     {
+        var jwt = Request.Headers["Authorization"];
+
+        if (!_authService.IsPersonId(jwt, request.PersonId))
+            return Forbid();
+
         var result = await _roomService.CreateAsync(request);
 
         if (result.Success)
@@ -124,6 +133,11 @@ public class RoomController : ControllerBase
     [HttpPut("update")]
     public async Task<ActionResult<OperationResult<int>>> Update(RoomUpdateRequest request)
     {
+        var jwt = Request.Headers["Authorization"];
+
+        if (!await _authService.CheckRoomPermissionAsync(jwt, request.Id, 2))
+            return Forbid();
+
         var result = await _roomService.UpdateAsync(request);
 
         if (result.Success)
@@ -140,6 +154,11 @@ public class RoomController : ControllerBase
     [HttpDelete("{id}")]
     public async Task<ActionResult<OperationResult<int>>> Delete(int id)
     {
+        var jwt = Request.Headers["Authorization"];
+
+        if (!await _authService.CheckRoomPermissionAsync(jwt, id, 1))
+            return Forbid();
+
         var result = await _roomService.DeleteAsync(id);
 
         if (result.Success)
